Update figures in the lfric datamodel docs (#49)
I've made an updated version of the eponymous PSyKAl diagram to fix some
issues with the currently included version. Also resized some of the
other figures to reduce the size difference between the text in the
figures and the docs. Finally use the tab feature of sphinx design to
neaten up the functionspace diagrams. The tab items required a zero
width space to allow Maths to be in the tab heading.